requirements.txt
kafka-python==2.0.2
avro-python3==1.10.2
pandas==2.2.2
pyarrow==16.1.0
python-schema-registry-client==2.5.6
pyspark==3.5.1
pymongo==4.7.2
delta-spark==3.1.0
# sqlalchemy==2.0.20
sqlalchemy==1.4.49
python-dotenv==1.0.0
psycopg2-binary==2.9.7
requests==2.31.0
numpy==1.26.4
minio==7.1.16
PyYAML==6.0.1
dbt-core==1.6.5
dbt-postgres==1.6.5
great_expectations==0.17.21
jupyterlab==4.0.6
apache-airflow==2.7.1
apache-airflow-providers-postgres==5.6.1
airflow-provider-great-expectations==0.2.6
apache-airflow-providers-docker==3.7.5
s3fs==0.4.2